added MPLearner to learner.py for mixed precision using torch.autocast and torch.cuda.amp.GradScalar
modified show images

modified Activation stats,in hist calculation .float() is used to work with mixed precision
Changes to be committed:
	modified:   init.py
	modified:   learner.py
	modified:   utils.py

Untracked files:
	.ipynb_checkpoints/
	__pycache__/

+++ miniai/init.py
@@ -18,7 +18,7 @@ class ActivationStats():
         out = to_cpu(out)
         h.stats[0].append(out.mean())
         h.stats[1].append(out.std())
-        h.stats[2].append(out.abs().histc(40,0,10))
+        h.stats[2].append(out.abs().float().histc(40,0,10))
     def get_dead_percentage(self,h):
         hist = torch.stack(h.stats[2])
         return hist[:,0]/hist.sum(1)
